Fold changes from the 6 Sept "branch" back onto the mainline.
The 6 Sept WD contains the ED->WD changes plus a while bunch of whitespace (typically at the end of line) cleanup. The 6 Sept version was the one that I used as a baseline when massaging the spec to use respec, so I need to check this version in so that we can have a series of clean diffs that are easy to review. The ED->WD change will need to be undone in a follow-up changelist.

Unfortunately, changes to top-of-tree after 6 Sept will also need to be re-applied. :-(  Thankfully, there are only a few that affect the DOM3-Events file (listed below).

Specifically, a follow-up changelist will need to fix:
* Change back WD -> ED for the top-of-tree version
* Change key name of Scroll to ScrollLock
* Fix spelling of mandatory (instead of manditory)