Generate key table from <key> tags.
This changelist replaces the current (manually coded) key table with a "keytable" that contains a single <key> tag for each key. The key table is generated from these tags.

There are no changes to the key table content in this changelist. This is to prepare for re-organizing this table into more manageable chunks.