Change initKeyboardEvent to be compatible with Blink/WebKit
Gecko is the only UA that doesn't implement initKeyboardEvent, and is
looking to implement it to improve interoperability.  However, the
parameter list in Blink and WebKit differs from the list in the spec,
which seemingly matches IE.  We would like to implement the Blink/WebKit
version because we believe it would be more compatible, so we would like
the spec to change to match it.  This will also require only Edge to
change its existing implementation instead of both Blink and WebKit, so
it will be easier to achieve interoperability.  (Assuming Edge's
implementation matches Microsoft documentation.)

The parameters are taken from Blink's implementation, except that the
first parameter is left mandatory although in Blink it's optional.
Gecko tried implementing the method with all parameters mandatory but
had to revert it because of incompatibility.  WebKit adds one more
parameter "altGraphKey" that I didn't put in the spec.

References:

  Microsoft docs: https://msdn.microsoft.com/en-us/library/ff975297(v=vs.85).aspx
  Blink IDL: https://src.chromium.org/viewvc/blink/trunk/Source/core/events/KeyboardEvent.idl?revision=198345#l43
  WebKit IDL: https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/dom/KeyboardEvent.idl?rev=209895#L52
  Mozilla bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1387828

I also updated the HTML40 and XFORMS references to HTML401 and XFORMS11
to silence bikeshed errors that they were obsolete.