Remove text selection from the dblclick default action - #328
Text selection begins on all browsers on the mousedown event (consistent
with OS text selection), allowing extending the selection by continuing
to drag. The dblclick event is too late to prevent text selection, so we
should remove this default action to prevent confusion.

Mousedown already includes text selection as a possible default action
in the note.