Remove definitions of MutationEvent and friends
Please see this WHATWG/dom issue for more details:

https://github.com/whatwg/dom/issues/305#issuecomment-2309207540

Mutation events have now been disabled for most all users of
the Chrome browser. While there are still a few sites opted
back in to the events via origin trial tokens, those are time
limited and we've committed to end-date them as soon as possible.
However, it feels appropriate to remove the definition of these
deprecated events from the standard at this point.