Fire keypress event for Enter, Shift+Enter, and Ctrl+Enter keys
For web compat and interop with actual behavior in browsers, this change
updates the spec to require firing “keypress” events for the Enter key
and for the combination of the Enter key with the Shift or Ctrl keys.

Otherwise, without this change, the spec is restricting implementations
to requirements for keypress that are different from what browsers need
to implement in order to be compatible with existing web content.